+++ NEWS.md
@@ -13,6 +13,8 @@
   cache GFF/GTF files when used in `makeGRangesFromGFF`.
 - Renamed `ignoreTxVersion` to simply `ignoreVersion`, where applicable.
   We want this setting to also apply at gene level.
+- No longer using `forceDetach` internally to force unload some Bioconductor
+  annotation packages.
 
 ## AcidGenomes 0.1.1 (2020-10-12)